pyproject.toml
[tool.poetry]
name = "qrcode-stl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
qrcode = "^7.3.1"
solidpython = "^1.1.3"
rq = "^1.10.1"
redis = "^4.2.0"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
